inputMode and bit mo stackframes
input mode on the vm lets you type in a script. parser fills in
stackframe vars inside bcParser.stackFrames with a bcStackFrameInfo
type, then copies it to bcFuncInfo if its a function. the generator also
creates the same stackframes in the bcExecContext at runtime, but
function calls arent done yet so they wont work properly